requirements.txt
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.28.0
PyQt6>=6.4.0
